[LINT] error 732
Loss of sign (Context) (Type to Type)

This mostly consists of a change to unsigned integers for any values
pertaining to sizes and indices. Sizes have been changed to size_t and indices
to idx_t (newly defined in stdafx.h).

However, the changes are rather invasive and I'm not 100% about their added
values. Should the maintainer of the project feel this commit isn't worthy of
the project, I'd understand if it's cherry picked out of the mainline branch.

+++ PythonScript/src/ProcessExecute.cpp
@@ -179,7 +179,7 @@ long ProcessExecute::execute(const TCHAR *commandLine, boost::python::object pyS
 
 		::ZeroMemory( &pi, sizeof(PROCESS_INFORMATION) );
 	
-		int commandLineLength = _tcslen(commandLine) + 1;
+		size_t commandLineLength = _tcslen(commandLine) + 1;
 		// We use an auto_ptr here because of a potential early out due to an exception thrown.
 		std::auto_ptr<TCHAR> cmdLine(new TCHAR[commandLineLength]);
 		_tcscpy_s(cmdLine.get(), commandLineLength, commandLine);
