+++ PythonScript/src/WcharMbcsConverter.cpp
@@ -73,7 +73,7 @@ std::shared_ptr<TCHAR> WcharMbcsConverter::char2tchar(const char* mbStr)
 #ifdef _UNICODE
 	return char2wchar(mbStr);
 #else
-	int len = strlen(mbStr) + 1;
+	size_t len = strlen(mbStr) + 1;
 	std::shared_ptr<TCHAR> result(new TCHAR[len]);
 	strcpy_s(result.get(), len, mbStr);
 	return result;
@@ -85,7 +85,7 @@ std::shared_ptr<char> WcharMbcsConverter::tchar2char(const TCHAR* tStr)
 #ifdef _UNICODE
 	return wchar2char(tStr);
 #else
-	int len = _tcslen(tStr) + 1;
+	size_t len = _tcslen(tStr) + 1;
 	std::shared_ptr<TCHAR> result(new TCHAR[len]);
 	strcpy_s(result.get(), len, tStr);
 	return result;
